convert js data files to json data files and implement the auto importing of the react-icons component based on the name stored in the data file
