+++ package.json
@@ -6,6 +6,7 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@headlessui/react": "^1.7.7",
+    "@loadable/component": "^5.15.3",
     "@mdx-js/mdx": "^2.3.0",
     "@mdx-js/react": "^2.3.0",
     "@mui/icons-material": "^5.11.0",
@@ -17,6 +18,7 @@
     "@testing-library/user-event": "^13.5.0",
     "@uiw/react-textarea-code-editor": "^2.0.6",
     "axios": "^1.2.1",
+    "babel-plugin-macros": "^3.1.0",
     "prism-react-renderer": "^1.3.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
