--- src/components/card/Card.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import './Card.css'
-
-function Cards({name}) {
-  return (
-    <div className='card'>
-        <h2>{name}</h2>
-    </div>
-  )
-}
-
-export default Cards
