--- src/components/card/ItemCard.scss
@@ -1,141 +1,0 @@
-@import '../../sass/abstract/variable';
-
-.item{
-    margin: 15px 0px;
-    padding: 10px;
-    border-radius: 6px;
-    display: flex;
-    
-    flex-direction: row;
-    align-items: center;
-    flex-direction: row;
-    background-color: white;
-    border: 1px solid black;
-    // width: 25.5rem; // results in unwanted horizontal scroll bar
-
-    @include mobile-l {
-        // width: 17.5rem;
-        flex-direction: row;
-    }
-
-    @include mobile-s {
-        width: 15rem;
-    }
-
-    img{
-        border-radius: 50%;
-        height: 100px;
-        width:100px;
-        margin:10px;
-    }
-    .detail{
-        @include mobile-l {
-            text-align: center;
-        }
-
-        // prevent long name to overflow box
-        h1{
-            
-            word-wrap: break-word;
-        }
-
-        a{
-            text-decoration: none;
-            color: white;
-            font-weight: 700;
-        }
-        button{
-            background-color: #439A97;
-            color: white;
-            padding: 5px 15px;
-            outline: none;
-            border: none;
-            border-radius: 6px;
-           @include mobile-l {
-                margin: 5% 0;
-           }
-        }
-
-        button:hover{
-            transform: scale(1.04);
-        }
-    }
-}
-
-.IndividualItemCard{
-    margin: 2rem;
-    padding: 1rem;
-    background-color:"#333";
-    width: 90%;
-}
-a{
-    text-decoration: none;
-    color: #333 ;
-    font-weight: 700;
-}
-
-.itemCardTitle{
-    display: flex;
-    
-    height: fit-content;
-    // width: 100%;
-}
-
-.itemCard2ndRow{
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-}
-
-.tag{
-    width: fit-content;
-    border: 1px solid black;
-    border-radius: 10px;
-    background:#333;
-    color: white;
-    padding: .2rem;
-    margin: .1rem;
-    
-}
-
-.tagWrapper{
-    margin-top:1rem;
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
-.bottomItemCard {
-    margin-top: 1rem;
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
-.Easy{
-    color:Green;
-}
-
-.originalItem{
-    margin-bottom: .4rem;
-}
-
-.difficultyItem, .originalItem{
-    position: relative;
-}
-
-.itemCardTitle:hover{
-    background:#232323;
-    color:aliceblue;
-    cursor:crosshair;
-}
-
-.itemCardTitle:hover .tag{
-      background:aliceblue;
-    color:#232323;
-  
-}
