+++ src/components/homepage/platformList/PlatformList.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PlatformCard from "../../card/PlatformCard";
+import PlatformCard from "../../LearningResources";
 
 //This is a component that appears on homepage and displays links leading to parts of site with content dedicated to different platforms or topics
 const PlatformList = () => {
