+++ src/pages/ItemPage.jsx
@@ -1,16 +1,13 @@
 import React, { useEffect } from "react";
-import { Container, Grid } from "@mui/material";
-import Footer from "../components/footer/Footer";
-import NavBar from "../components/navbar/Navbar";
 import ReactPlayer from 'react-player'
-import { HiVideoCamera } from 'react-icons/hi'
+import Footer from "../components/Footer";
+import NavBar from "../components/Navbar";
+import Coder from "../components/Coder";
 import { BsFillImageFill, BsFilm } from 'react-icons/bs'
-import { TbMovie } from 'react-icons/tb'
-import CodeEditor, { SelectionText } from "@uiw/react-textarea-code-editor";
+import { SelectionText } from "@uiw/react-textarea-code-editor";
 import { useParams } from 'react-router-dom';
 import { LeetCodeSolutions } from "../data/solutionLC";
 import { CodeWarsSolutions } from "../data/solutionCW";
-import HighlightComponent from "../components/highlight/HighlightComponent";
 
 export default function ItemPage({ sol }) {
 
@@ -69,7 +66,7 @@ export default function ItemPage({ sol }) {
           </div>
           <div>
             <h1 className="text-2xl">Javascript Solution</h1>
-            <HighlightComponent code={javascript} />
+            <Coder code={javascript} />
           </div>
           <div className="mb-12 min-h-500">
             <h1 className="text-2xl">Video Solution</h1>
